install swl, cleacve and upload image VDO 10

+++ resources/js/Pages/Store.vue
@@ -11,7 +11,13 @@
             <button type="button" @click="CancelStore()" class="btn rounded-pill btn-danger">ຍົກເລີກ</button>
         </div>
         <div class="row">
-            <div class="col-md-3">Image</div>
+            <div class="col-md-3 d-flex justify-content-center" style="position:relative;">
+                <button type="button" v-if="FormStore.image" @click="RemoveImg()" class="btn rounded-pill btn-icon btn-danger img-x">
+                    <i class='bx bx-x fs-3'></i>
+               </button>
+                <img :src="Image_preview" @click="$refs.img_store.click()" class="img-store cursor-pointer "  >
+                <input type="file" ref="img_store" style=" display:none; " @change="onSelect($event)" >
+            </div>
             <div class="col-md-9">
                 <div class="mb-2">
                     <label  class="form-label fs-6">ຊື່ສິນຄ້າ:</label>
@@ -19,13 +25,13 @@
                 </div>
                 <div class="mb-2">
                     <label  class="form-label fs-6">ຈຳນວນ:</label>
-                    <input type="text" class="form-control" v-model="FormStore.qty"  placeholder="..." >
+                    <cleave :options="options" class="form-control" v-model="FormStore.qty"  placeholder="..." />
                 </div>
                 <div class="row">
                     <div class="col-md-6">
                         <label  class="form-label fs-6">ລາຄາຊື້:</label>
                         <div class="input-group">
-                            <input type="text" class="form-control" v-model="FormStore.price_buy" placeholder="..." >
+                            <cleave :options="options" class="form-control text-end" v-model="FormStore.price_buy" placeholder="..." />
                             <span class="input-group-text" >ກີບ</span>
                         </div>
 
@@ -33,7 +39,7 @@
                     <div class="col-md-6">
                         <label  class="form-label fs-6">ລາຄາຂາຍ:</label>
                         <div class="input-group">
-                            <input type="text" class="form-control" v-model="FormStore.price_sell" placeholder="..." >
+                            <cleave :options="options" class="form-control text-end" v-model="FormStore.price_sell" placeholder="..." />
                             <span class="input-group-text" >ກີບ</span>
                         </div>
                     </div>
@@ -89,7 +95,11 @@
             <tbody>
             <tr v-for="item in StoreData.data" :key="item">
                 <td class="text-center">{{ item.id }}</td>
-                <td>{{  }} </td>
+                <td>
+                    <img v-if="item.image" :src="url + '/assets/img/'+ item.image "  class="list-img border-1 p-1 rounded shadow-sm">
+                    <img v-else :src="url + '/assets/img/no_img.jpg' " class="list-img border-1 p-1 rounded shadow-sm">
+                
+                </td>
                 <td>{{ item.name }}</td>
                 <td class="text-center">{{ formatPrice(item.qty) }}</td>
                 <td class="text-end">{{ formatPrice(item.price_buy) }}</td>
@@ -105,6 +115,8 @@
             </tr>
             </tbody>
         </table>
+
+        <button class="btn btn-primary" @click="showAlert">Hello world</button>
 
 
         <Pagination :pagination="StoreData" :offset="4" @paginate='GetStore($event)'  />
@@ -127,6 +139,8 @@ export default {
     }, 
     data() {
         return {
+            url: window.location.origin,
+            Image_preview: window.location.origin + '/assets/img/upload_img.jpg',
             ShowForm:false,
             FormType:true,
             FormStore:{
@@ -141,6 +155,16 @@ export default {
             PerPage:5,
             Search:'',
             EditID:'',
+            options: {
+                  numeral: true,
+                  numeralPositiveOnly: true,
+                  noImmediatePrefix: true,
+                  rawValueTrimPrefix: true,
+                  numeralIntegerScale: 10,
+                  numeralDecimalScale: 2,
+                  numeralDecimalMark: '.',
+                  delimiter: ','
+                }
         }
     },
     computed:{
@@ -153,9 +177,61 @@ export default {
         }
     },
     methods:{
+        RemoveImg(){
+            this.FormStore.image = '';
+            this.Image_preview = window.location.origin + '/assets/img/upload_img.jpg';
+        },
+        showAlert() {
+            // Use sweetalert2
+            // sussecc
+            // this.$swal({
+            //     toast: true,
+            //     position: "top-end",
+            //     icon: "success",
+            //     title: "Your work has been saved",
+            //     showConfirmButton: false,
+            //     timer: 1500
+            //     });
+
+                // error
+                // this.$swal({
+                //     icon: "error",
+                //     title: "ຜິດຜາດ!",
+                //     text: "That thing is still around?",
+                //     showConfirmButton: false,
+                //     timer: 3500
+                // });
+                this.$swal({
+                        title: "ທ່ານແນ່ໃຊບໍ່?",
+                        text: "ທີ່ຂະລຶຍຂໍ້ມູນນູນນີ້!",
+                        icon: "warning",
+                        showCancelButton: true,
+                        confirmButtonColor: "#3085d6",
+                        cancelButtonColor: "#d33",
+                        confirmButtonText: "ຕົກລົງ!",
+                        cancelButtonText: "ຍົກເລີກ!",
+                        }).then((result) => {
+                        if (result.isConfirmed) {
+                            this.$swal({
+                            title: "ການລຶບຂໍ້ມູນ!",
+                            text: "res.data.message",
+                            icon: "success"
+                            });
+                        }
+                        })
+            },
         formatPrice(value) {
             let val = (value / 1).toFixed(0).replace(",", ".");
             return val.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+        },
+        onSelect(event){
+            // console.log(event);
+                this.FormStore.image = event.target.files[0];
+                let reader = new FileReader();
+                reader.readAsDataURL(this.FormStore.image);
+                reader.addEventListener("load", function(){
+                    this.Image_preview = reader.result
+                }.bind(this,false));
         },
         AddStore(){
             this.ShowForm = true
@@ -178,6 +254,8 @@ export default {
             this.EditID = id;
             // ປ່ຽນສະຖານະ ຟອມເປັນການອັບເດດ
             this.FormType = false;
+
+            
             
             axios.get(`api/store/edit/${id}`,{ headers:{ Authorization: 'Bearer '+this.store.get_token }}).then((res)=>{
 
@@ -205,42 +283,94 @@ export default {
         },
         DelStore(id){
 
-            axios.delete(`api/store/delete/${id}`,{ headers:{ Authorization: 'Bearer '+this.store.get_token }}).then((res)=>{
+            this.$swal({
+                        title: "ທ່ານແນ່ໃຊບໍ່?",
+                        text: "ທີ່ຂະລຶຍຂໍ້ມູນນູນນີ້!",
+                        icon: "warning",
+                        showCancelButton: true,
+                        confirmButtonColor: "#3085d6",
+                        cancelButtonColor: "#d33",
+                        confirmButtonText: "ຕົກລົງ!",
+                        cancelButtonText: "ຍົກເລີກ!",
+                        }).then((result) => {
+                            if (result.isConfirmed) {
+                            axios.delete(`api/store/delete/${id}`,{ headers:{ Authorization: 'Bearer '+this.store.get_token }}).then((res)=>{
 
-                if(res.data.success){
-                    // get store
-                    this.GetStore();
-                } else {
+                                if(res.data.success){
 
-                }
+                       
 
-            }).catch((error)=>{
-                if(error.response){
-                        if(error.response.status==401){
-                            // ເຄຼຍຂໍ້ມູນໃນ localstorage
-                            localStorage.removeItem('web_token');
-                            localStorage.removeItem('web_user');
+                                        this.$swal({
+                                                title: "ການລຶບຂໍ້ມູນ!",
+                                                text: res.data.message,
+                                                icon: "success"
+                                        });
 
-                            // ເຄຼຍຂໍ້ມູນໃນ pinia
-                            this.store.remove_token();
-                            this.store.remove_user();
 
-                            // go to login page
-                            this.$router.push('/login');
+                                                // get store
+                                            this.GetStore();
+
+
+                                        } else {
+
+                                                this.$swal({
+                                                icon: "error",
+                                                title: "ຜິດຜາດ!",
+                                                text: res.data.message,
+                                                showConfirmButton: false,
+                                                timer: 3500
+                                                });
+                                    }
+                        
+                            });
                         }
-                    }
-            })
+                
+                    }).catch((error)=>{
+                        if(error.response){
+                                if(error.response.status==401){
+                                    // ເຄຼຍຂໍ້ມູນໃນ localstorage
+                                    localStorage.removeItem('web_token');
+                                    localStorage.removeItem('web_user');
 
+                                    // ເຄຼຍຂໍ້ມູນໃນ pinia
+                                    this.store.remove_token();
+                                    this.store.remove_user();
+
+                                    // go to login page
+                                    this.$router.push('/login');
+                                }
+                            }
+                    });
+
+                    
 
         },
         SaveStore(){
             if(this.FormType){
                 // ເພີ່ມຂໍ້ມູນໃໝ່
-                axios.post('api/store/add',this.FormStore, { headers:{ Authorization: 'Bearer '+this.store.get_token }}).then((res)=>{
+                axios.post('api/store/add',this.FormStore, { headers:{ "Content-Type":"multipart/form-data", Authorization: 'Bearer '+this.store.get_token }}).then((res)=>{
                     if(res.data.success){
                         this.ShowForm = false;
                         this.GetStore();
+
+                        this.$swal({
+                            toast: true,
+                            position: "top-end",
+                            icon: "success",
+                            title: res.data.message,
+                            showConfirmButton: false,
+                            timer: 1500
+                        });
+
                     } else {
+
+                        this.$swal({
+                            icon: "error",
+                            title: "ຜິດຜາດ!",
+                            text: res.data.message,
+                            showConfirmButton: false,
+                            timer: 3500
+                        });
 
                     }
                 }).catch((error)=>{
@@ -264,13 +394,29 @@ export default {
             } else {
                 // ອັບເດດຂໍ້ມູນ
                 // update store
-                axios.post(`api/store/update/${this.EditID}`,this.FormStore, { headers:{ Authorization: 'Bearer '+this.store.get_token }}).then((res)=>{
+                axios.post(`api/store/update/${this.EditID}`,this.FormStore, { headers:{ "Content-Type":"multipart/form-data", Authorization: 'Bearer '+this.store.get_token }}).then((res)=>{
 
                     if(res.data.success){
                         this.ShowForm = false;
                         this.GetStore();
-                    } else {
+                        this.$swal({
+                            toast: true,
+                            position: "top-end",
+                            icon: "success",
+                            title: res.data.message,
+                            showConfirmButton: false,
+                            timer: 1500
+                        });
 
+                    } else {
+                        
+                        this.$swal({
+                            icon: "error",
+                            title: "ຜິດຜາດ!",
+                            text: res.data.message,
+                            showConfirmButton: false,
+                            timer: 3500
+                        });
                     }
 
                 }).catch((error)=>{
@@ -355,6 +501,21 @@ export default {
     }
 }
 </script>
-<style lang="">
-    
+<style scoped>
+    .img-store{
+        width: 80%;
+        height: auto;
+        object-fit: contain;
+    }
+    .img-x{
+        position: absolute;
+        top: 0px;
+        right: 5px;
+    }
+    .list-img{
+        width: 80px;
+        height: 80px;
+        object-fit: cover;
+        object-position: center;
+    }
 </style>
